Add PaperCharacter implementation and refine character APIs
Introduced `PaperCharacter` and `PaperPlayerCharacter` classes to enhance character handling, aligning with Bukkit's entity model. Updated APIs to use UUIDs for viewer tracking, replaced `short` with `byte` for skin parts, and added a `SkinPartBuilder` factory method for improved usability. These changes streamline NPC management and ensure better integration within the framework.

+++ src/main/java/net/thenextlvl/character/model/CharacterSkinPartBuilder.java
@@ -6,7 +6,7 @@ import org.jspecify.annotations.NullMarked;
 
 @NullMarked
 public class CharacterSkinPartBuilder implements SkinPartBuilder {
-    private short raw;
+    private byte raw;
 
     public CharacterSkinPartBuilder() {
         this((byte) 127);
